// src/app/model/bean/archivo-dpmn.model.ts
export class ArchivoDpmn {
  id: number;
  numCorrelativoDpmn: number;
  numeroDpmn: string;
  codAduanaDpmn: string;
  annioDpmn: string;
  codTipoDocumento: string;
  desTipoDocumento: string;
  origenInvocacion: string;
  usuarioRegistra: string;
  nomArchivo: string;
  nomContentType: string;
  valArchivoBase64: string;
  fechaRegistro: Date;
}